delete the old file
